// sah-ai/lib/types.ts
export interface Facility {
  readonly name: string;
  readonly address: string;
  readonly city: string;
  readonly state: string;
  readonly zip: string;
  readonly phone: string;
  readonly website: string;
  readonly latitude: number | null;
  readonly longitude: number | null;
  readonly services: readonly string[];
  readonly payment: readonly string[];
  readonly specialPrograms: readonly string[];
  readonly languages: readonly string[];
  readonly facilityType: string;
}